captchaResolver/requirements.txt
Flask==2.0.3
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
numpy==1.24.3
tensorflow==2.13.1
Pillow==9.0.1
keras==2.13.1
opencv-python==4.8.1.78
Werkzeug==2.2.2